timeout signature for fsockopen is float

+++ src/Connection/UdpSocket.php
@@ -58,10 +58,10 @@ class UdpSocket extends InetSocket implements Connection
     /**
      * @param string $host
      * @param int $port
-     * @param int $timeout
+     * @param float $timeout
      * @param bool $persistent
      */
-    protected function connect(string $host, int $port, int $timeout, bool $persistent = false): void
+    protected function connect(string $host, int $port, float $timeout, bool $persistent = false): void
     {
         $errorNumber = 0;
         $errorMessage = '';
